requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
Naked==0.1.31
pycryptodome==3.10.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
urllib3==1.26.5
